fixed bug that caused infinite rerender
i have come to realize that useEffect really only needs to be called when mounting the component and that is it. I have yet to find a reason why it would be called at any other time. As long as you change the state of the field, that will cause the re-render achieving exactly what I am looking for.